+++ lib/DB/GnfConnectionProvider.php
@@ -12,7 +12,7 @@ use Ridibooks\Platform\Common\Cache\AdaptableCache;
 class GnfConnectionProvider
 {
     /**
-     * @var \Doctrine\DBAL\Connection[]
+     * @var base[]
      */
     private static $connection_pool = [];
 
@@ -25,34 +25,10 @@ class GnfConnectionProvider
     public static function getConnection($group_name)
     {
         if (!isset(self::$connection_pool[$group_name])) {
-            self::$connection_pool[$group_name] = self::createConnection($group_name);
+            self::$connection_pool[$group_name] = new PDO(self::createConnection($group_name));
         }
 
-        return new PDO(self::$connection_pool[$group_name]);
-    }
-
-    /**
-     * @param      $group_name
-     * @param bool $is_auto_reconnect
-     *
-     * @return Connection
-     * @throws \Doctrine\DBAL\DBALException
-     */
-    public static function getConnectionWithAutoReconnection($group_name, $is_auto_reconnect = false)
-    {
-        if (!isset(self::$connection_pool[$group_name])) {
-            self::$connection_pool[$group_name] = self::createConnection($group_name);
-        }
-
-        $connection = self::$connection_pool[$group_name];
-
-        // 접속 보장을 해야하는 경우 핑을 확인
-        if ($is_auto_reconnect && $connection->ping() === false) {
-            $connection->close();
-            $connection->connect();
-        }
-
-        return $connection;
+        return self::$connection_pool[$group_name];
     }
 
     /**