+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2019-09-18
+### Changed
+- Change Variable Type of `connection_pool` from `Doctrine/Connection` to `PDO base` for compatibility of usage
+### Removed
+- Remove Function `getConnectionWithAutoReconnection` in `GnfConnectionProvider`
+
 ## [0.8.1] - 2019-09-17
 ### Added
 - Added Function `renameFile` in `FileUtils`